Move AttentionOpWithRNG to op
